PRSD-945: Moves MaximumLengthInputStream into helper folder
